requirements.txt
torch
nltk
python-Levenshtein 